kvsserver: split code into modules
